DEFENSE_5vs0.txt is 2 Marks, 1 SmartGoalie, 1 bot of a TwoManWall, and 1 Field(Clear)Bot for bottom of field. OFFENSE_5vs0.txt is 1 Striker, 1 Assist, 1 SmartGoalie, and 2 Field(Clear)Bots for top and bottom of field. numrobotsincircle works.

+++ strategy/Core Play Files/Functions.cs
@@ -485,11 +485,11 @@ namespace Robocup.Plays
                             return new Vector2(p1.X + v.X, p1.Y + v.Y);
                         });
             addFunction("numberOfRobotsInACircle", "Circle, Team - Number of Robots",
-                        "In a circle ~, the number of Robots of the Team ~ in it", typeof(int),
+                        "In a circle ~, the number of Robots of the Team ~ in it", typeof(double),
                         new Type[] { typeof(Circle), typeof(TeamCondition) },
                         delegate(EvaluatorState state, object[] objects)
                         {
-                            int count = 0;
+                            double count = 0;
                             TeamCondition condition = (TeamCondition)objects[1];
                             List<RobotInfo> allinfos = new List<RobotInfo>();
                             if (condition.maybeOurs())
